codelabs/trusted_space_codelab/src/client/requirements.txt
pyjwt
requests
cryptography
google-cloud-kms==2.21.1